+++ packages/common/src/utils/indent.ts
@@ -1,0 +1,3 @@
+export function indent(message: string, indentation = "  "): string {
+  return message.replaceAll(/(^|\n)/g, `$1${indentation}`);
+}