+++ packages/common/src/writeContract.ts
@@ -9,13 +9,8 @@ import {
   ContractFunctionName,
   ContractFunctionArgs,
   PublicClient,
-  encodeFunctionData,
-  EncodeFunctionDataParameters,
 } from "viem";
-import {
-  prepareTransactionRequest as viem_prepareTransactionRequest,
-  writeContract as viem_writeContract,
-} from "viem/actions";
+import { writeContract as viem_writeContract } from "viem/actions";
 import pRetry from "p-retry";
 import { debug as parentDebug } from "./debug";
 import { getNonceManager } from "./getNonceManager";
@@ -46,7 +41,7 @@ export type WriteContractExtraOptions<chain extends Chain | undefined> = {
 export async function writeContract<
   chain extends Chain | undefined,
   account extends Account | undefined,
-  abi extends Abi | readonly unknown[],
+  const abi extends Abi | readonly unknown[],
   functionName extends ContractFunctionName<abi, "nonpayable" | "payable">,
   args extends ContractFunctionArgs<abi, "nonpayable" | "payable", functionName>,
   chainOverride extends Chain | undefined,
@@ -63,15 +58,11 @@ export async function writeContract<
   const account = parseAccount(rawAccount);
   const chain = client.chain;
 
-  const defaultParameters = {
-    chain,
-    ...(chain?.fees ? { type: "eip1559" } : {}),
-  } satisfies Omit<WriteContractParameters, "address" | "abi" | "account" | "functionName">;
-
+  const blockTag = "pending";
   const nonceManager = await getNonceManager({
     client: opts.publicClient ?? client,
     address: account.address,
-    blockTag: "pending",
+    blockTag,
     queueConcurrency: opts.queueConcurrency,
   });
 
@@ -81,73 +72,32 @@ export async function writeContract<
     args: { chain },
   });
 
-  async function prepare(): Promise<WriteContractParameters<abi, functionName, args, chain, account, chainOverride>> {
-    if (request.gas) {
-      debug("gas provided, skipping preparation", request.functionName, request.address);
-      return request;
-    }
-
-    const { abi, address, args, dataSuffix, functionName } = request;
-    const data = encodeFunctionData({
-      abi,
-      args,
-      functionName,
-    } as EncodeFunctionDataParameters);
-
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { nonce, maxFeePerGas, maxPriorityFeePerGas, ...preparedTransaction } = await getAction(
-      client,
-      viem_prepareTransactionRequest,
-      "prepareTransactionRequest",
-    )({
-      // The fee values don't need to be accurate for gas estimation
-      // and we can save a couple rpc calls by providing stubs here.
-      // These are later overridden with accurate values from `feeRef`.
-      maxFeePerGas: 0n,
-      maxPriorityFeePerGas: 0n,
-      // Send the current nonce without increasing the stored value
-      nonce: nonceManager.getNonce(),
-      ...defaultParameters,
-      ...request,
-      blockTag: "pending",
-      account,
-      // From `viem/writeContract`
-      data: `${data}${dataSuffix ? dataSuffix.replace("0x", "") : ""}`,
-      to: address,
-    } as never);
-
-    return preparedTransaction as never;
-  }
-
   return nonceManager.mempoolQueue.add(
     () =>
       pRetry(
         async () => {
-          if (!nonceManager.hasNonce()) {
-            await nonceManager.resetNonce();
-          }
-
-          // We estimate gas before increasing the local nonce to prevent nonce gaps.
-          // Invalid transactions fail the gas estimation step are never submitted
-          // to the network, so they should not increase the nonce.
-          const preparedRequest = await prepare();
-
           const nonce = nonceManager.nextNonce();
-
-          const fullRequest = { ...preparedRequest, nonce, ...feeRef.fees };
-          debug("calling", fullRequest.functionName, "with nonce", nonce, "at", fullRequest.address);
-          return await viem_writeContract(client, fullRequest as never);
+          const params = {
+            blockTag,
+            ...request,
+            nonce,
+            ...feeRef.fees,
+          } as const satisfies WriteContractParameters<abi, functionName, args, chain, account, chainOverride>;
+          debug("calling", params.functionName, "at", params.address, "with nonce", nonce);
+          return await getAction(client, viem_writeContract, "writeContract")(params as never);
         },
         {
           retries: 3,
           onFailedAttempt: async (error) => {
-            // On nonce errors, reset the nonce and retry
+            // in case this tx failed before hitting the mempool (i.e. gas estimation error), reset nonce so we don't skip past the unused nonce
+            debug("failed, resetting nonce");
+            await nonceManager.resetNonce();
+            // retry nonce errors
+            // TODO: upgrade p-retry and move this to shouldRetry
             if (nonceManager.shouldResetNonce(error)) {
               debug("got nonce error, retrying", error.message);
-              await nonceManager.resetNonce();
               return;
             }
-            // TODO: prepareWrite again if there are gas errors?
             throw error;
           },
         },