+++ packages/store/ts/config/v2/table.ts
@@ -47,7 +47,7 @@ export type ValidateTableOptions = { inStoreContext: boolean };
 
 export type requiredTableKey<inStoreContext extends boolean> = Exclude<
   requiredKeyOf<TableInput>,
-  inStoreContext extends true ? "label" | "namespace" : ""
+  inStoreContext extends true ? "label" | "namespaceLabel" | "namespace" : never
 >;
 
 export type validateTable<
@@ -59,9 +59,9 @@ export type validateTable<
     ? validateKeys<getStaticAbiTypeKeys<conform<get<input, "schema">, SchemaInput>, scope>, get<input, key>>
     : key extends "schema"
       ? validateSchema<get<input, key>, scope>
-      : key extends "label" | "namespace"
+      : key extends "label" | "namespaceLabel" | "namespace"
         ? options["inStoreContext"] extends true
-          ? ErrorMessage<"Overrides of `label` and `namespace` are not allowed for tables in this context">
+          ? ErrorMessage<"Overrides of `label`, `namespaceLabel`, and `namespace` are not allowed for tables in this context">
           : key extends keyof input
             ? narrow<input[key]>
             : never
@@ -99,12 +99,29 @@ export function validateTable<input, scope extends Scope = AbiTypeScope>(
   if (hasOwnKey(input, "namespace") && typeof input.namespace === "string" && input.namespace.length > 14) {
     throw new Error(`Table \`namespace\` must fit into a \`bytes14\`, but "${input.namespace}" is too long.`);
   }
+
+  if (
+    hasOwnKey(input, "namespaceLabel") &&
+    typeof input.namespaceLabel === "string" &&
+    (!hasOwnKey(input, "namespace") || typeof input.namespace !== "string") &&
+    input.namespaceLabel.length > 14
+  ) {
+    throw new Error(
+      `Table \`namespace\` defaults to \`namespaceLabel\`, but must fit into a \`bytes14\` and "${input.namespaceLabel}" is too long. Provide explicit \`namespace\` override.`,
+    );
+  }
+
   if (hasOwnKey(input, "name") && typeof input.name === "string" && input.name.length > 16) {
     throw new Error(`Table \`name\` must fit into a \`bytes16\`, but "${input.name}" is too long.`);
   }
 
-  if (options.inStoreContext && (hasOwnKey(input, "label") || hasOwnKey(input, "namespace"))) {
-    throw new Error("Overrides of `label` and `namespace` are not allowed for tables in this context.");
+  if (
+    options.inStoreContext &&
+    (hasOwnKey(input, "label") || hasOwnKey(input, "namespaceLabel") || hasOwnKey(input, "namespace"))
+  ) {
+    throw new Error(
+      "Overrides of `label`, `namespaceLabel`, and `namespace` are not allowed for tables in this context.",
+    );
   }
 }
 
@@ -139,6 +156,9 @@ export function resolveTableCodegen<input extends TableInput>(input: input): res
 export type resolveTable<input, scope extends Scope = Scope> = input extends TableInput
   ? {
       readonly label: input["label"];
+      readonly namespaceLabel: undefined extends input["namespaceLabel"]
+        ? typeof TABLE_DEFAULTS.namespaceLabel
+        : input["namespaceLabel"];
       readonly type: undefined extends input["type"] ? typeof TABLE_DEFAULTS.type : input["type"];
       readonly namespace: string;
       readonly name: string;
@@ -156,16 +176,20 @@ export function resolveTable<input extends TableInput, scope extends Scope = Abi
   input: input,
   scope: scope = AbiTypeScope as unknown as scope,
 ): resolveTable<input, scope> {
+  const namespaceLabel = input.namespaceLabel ?? TABLE_DEFAULTS.namespaceLabel;
+  // validate ensures this is length constrained
+  const namespace = input.namespace ?? namespaceLabel;
+
   const label = input.label;
-  const type = input.type ?? TABLE_DEFAULTS.type;
-  const namespace = input.namespace ?? TABLE_DEFAULTS.namespace;
   const name = input.name ?? label.slice(0, 16);
+  const type = input.type ?? TABLE_DEFAULTS.type;
   const tableId = resourceToHex({ type, namespace, name });
 
   return {
     label,
     type,
     namespace,
+    namespaceLabel,
     name,
     tableId,
     schema: resolveSchema(input.schema, scope),