+++ packages/abi-ts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @latticexyz/abi-ts
 
+## 2.2.1
+
+## 2.2.0
+
+## 2.1.1
+
 ## 2.1.0
 
 ### Patch Changes