+++ packages/react/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log
 
+## 2.2.1
+
+### Patch Changes
+
+- @latticexyz/store@2.2.1
+- @latticexyz/recs@2.2.1
+
+## 2.2.0
+
+### Patch Changes
+
+- Updated dependencies [04c675c]
+  - @latticexyz/store@2.2.0
+  - @latticexyz/recs@2.2.0
+
+## 2.1.1
+
+### Patch Changes
+
+- Updated dependencies [9e21e42]
+- Updated dependencies [57bf8c3]
+  - @latticexyz/store@2.1.1
+  - @latticexyz/recs@2.1.1
+
 ## 2.1.0
 
 ### Patch Changes