+++ packages/recs/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change Log
 
+## 2.2.1
+
+### Patch Changes
+
+- @latticexyz/schema-type@2.2.1
+- @latticexyz/utils@2.2.1
+
+## 2.2.0
+
+### Patch Changes
+
+- @latticexyz/schema-type@2.2.0
+- @latticexyz/utils@2.2.0
+
+## 2.1.1
+
+### Patch Changes
+
+- Updated dependencies [9e21e42]
+  - @latticexyz/schema-type@2.1.1
+  - @latticexyz/utils@2.1.1
+
 ## 2.1.0
 
 ### Patch Changes