+++ packages/store-indexer/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @latticexyz/store-indexer
 
+## 2.2.1
+
+### Patch Changes
+
+- Updated dependencies [603b2ab]
+- Updated dependencies [c0764a5]
+  - @latticexyz/store-sync@2.2.1
+  - @latticexyz/common@2.2.1
+  - @latticexyz/block-logs-stream@2.2.1
+  - @latticexyz/protocol-parser@2.2.1
+  - @latticexyz/store@2.2.1
+
+## 2.2.0
+
+### Patch Changes
+
+- Updated dependencies [69cd0a1]
+- Updated dependencies [04c675c]
+  - @latticexyz/common@2.2.0
+  - @latticexyz/store@2.2.0
+  - @latticexyz/block-logs-stream@2.2.0
+  - @latticexyz/protocol-parser@2.2.0
+  - @latticexyz/store-sync@2.2.0
+
+## 2.1.1
+
+### Patch Changes
+
+- 6435481: Upgrade `zod` to `3.23.8` to avoid issues with [excessively deep type instantiations](https://github.com/colinhacks/zod/issues/577).
+- 9e21e42: Bumped viem to `2.19.8` and abitype to `1.0.5`.
+
+  MUD projects using viem or abitype should do the same to ensure no type errors due to mismatched versions:
+
+  ```
+  pnpm recursive up viem@2.19.8 abitype@1.0.5
+  ```
+
+- Updated dependencies [6435481]
+- Updated dependencies [9e21e42]
+- Updated dependencies [2daaab1]
+- Updated dependencies [57bf8c3]
+  - @latticexyz/store-sync@2.1.1
+  - @latticexyz/block-logs-stream@2.1.1
+  - @latticexyz/common@2.1.1
+  - @latticexyz/protocol-parser@2.1.1
+  - @latticexyz/store@2.1.1
+
 ## 2.1.0
 
 ### Patch Changes